Update movie.py
Fixes a bug where titles with parens would return the value within the first set of parens as the year.

Now will look for the last set of parens, check if it's an int, and then return as year. Still possible to get bad data if a movie title ends with a positive int in parens AND has no year.

If no valid year is found, year is set to None.